Add files via upload
First, it was necessary to filter the number of subjects, as the database did not have the same number of individuals for each state. I created another dataset with a more restricted yet still representative sample (300 subjects). Then, I immediately analyzed the relationship between heart attack risk and state name. It seems that Mizoram has something different compared to other states, with a higher risk. It would be interesting to understand the cause, as there is no correlation with the variables available in the dataset. Heart attacks are more correlated with triglyceride levels in Goa, although it does not have the highest triglyceride levels; that distinction belongs to Haryana. However, there seems to be a correlation between the variables. Goa has the highest correlation between heart attacks, diabetes, and triglycerides. Notably:
1. Diabetes is more common or has higher levels in Goa.
2. Triglyceride levels are higher in Goa.
3. There is a strong correlation between triglycerides and heart attack risk in Goa.
These combined factors suggest that Goa has a higher risk of heart attacks, particularly due to high diabetes and triglyceride levels. Haryana has the highest average triglyceride levels. Mizoram has the highest average heart attack risk. Uttarakhand has the highest average stress levels.
Age and stress are only slightly correlated. On average, Manipur and Bihar have higher incomes. In terms of healthcare access (binary), Chhattisgarh has the highest values, but Bihar also has high values, which is reflected in the higher incomes we observed earlier.